pyproject.toml
[project]
name = "w-medcat"
version = "0.1.0"
description = "MedCAT NER annotation worker for NeuLANDER using FastStream, RabbitMQ and Azure Blob Storage"
readme = "README.md"
requires-python = "~=3.11"
dependencies = [
    "faststream[cli,rabbit]>=0.5.33",
    "medcat==1.15.0b",
    "neulander-core>=0.1.0",
    "python-dotenv>=1.0.1",
    "torch>=2.5.1,<3.0.0",
    "uvicorn>=0.34.0",
]

[dependency-groups]
dev = ["pre-commit>=4.0.1", "watchfiles>=1.0.3"]

[tool.uv.sources]
neulander-core = { path = '../core/' }
